Sonar fixes and code formatting

+++ cepaslib/src/main/java/com/itachi1706/cepaslib/card/cepas/CEPASProtocol.java
@@ -26,8 +26,9 @@
 package com.itachi1706.cepaslib.card.cepas;
 
 import android.nfc.tech.IsoDep;
-import androidx.annotation.NonNull;
 import android.util.Log;
+
+import androidx.annotation.NonNull;
 
 import com.itachi1706.cepaslib.card.cepas.raw.RawCEPASHistory;
 import com.itachi1706.cepaslib.card.cepas.raw.RawCEPASPurse;
@@ -47,7 +48,8 @@ class CEPASProtocol {
     private static final byte OPERATION_OK = (byte) 0x00;
     private static final byte PERMISSION_DENIED = (byte) 0x9D;
 
-    @NonNull private final IsoDep mTagTech;
+    @NonNull
+    private final IsoDep mTagTech;
 
     CEPASProtocol(@NonNull IsoDep tagTech) {
         mTagTech = tagTech;
@@ -125,7 +127,7 @@ class CEPASProtocol {
             }
 
             throw new CEPASException("Got generic invalid response: "
-                    + Integer.toHexString(((int) recvBuffer[recvBuffer.length - 2]) & 0xff));
+                    + Integer.toHexString((recvBuffer[recvBuffer.length - 2]) & 0xff));
         }
 
         output.write(recvBuffer, 0, recvBuffer.length - 2);
